Move location detail from unsorted
