+++ modernomad/core/views/unsorted.py
@@ -58,33 +58,6 @@ from bank.models import Account, Currency, Transaction, Entry
 logger = logging.getLogger(__name__)
 
 
-def location_detail(request, location_slug):
-    try:
-        location = (
-            Location.objects
-            .prefetch_related('house_admins', 'resources')
-            .get(slug=location_slug)
-        )
-        logger.debug(location.get_members())
-        logger.debug('--')
-        logger.debug(request.user)
-
-        if location.visibility == 'public' or location.visibility == 'link':
-            has_permission = True
-        elif request.user in location.get_members():
-            has_permission = True
-        else:
-            has_permission = False
-
-        if not has_permission:
-            raise Location.DoesNotExist
-
-    except Location.DoesNotExist:
-        raise Http404("The location does not exist or you do not have permission to view it")
-
-    return render(request, "location_detail.html", {'location': location, 'max_days': location.max_booking_days})
-
-
 def community(request, location_slug):
     location = get_object_or_404(Location, slug=location_slug)
     residents = location.residents()
